Review updates

* Move common factorization code to factorization_kernels
* Fix typos
* Add missing whitespace
