+++ cuda/factorization/par_ilu_kernels.cu
@@ -33,19 +33,12 @@ OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 #include "core/factorization/par_ilu_kernels.hpp"
 
 
-#include <ginkgo/core/base/array.hpp>
 #include <ginkgo/core/base/std_extensions.hpp>
 #include <ginkgo/core/matrix/coo.hpp>
 
 
-#include "core/components/prefix_sum.hpp"
-#include "core/matrix/csr_builder.hpp"
-#include "cuda/base/config.hpp"
 #include "cuda/base/math.hpp"
 #include "cuda/base/types.hpp"
-#include "cuda/components/cooperative_groups.cuh"
-#include "cuda/components/intrinsics.cuh"
-#include "cuda/components/searching.cuh"
 #include "cuda/components/thread_ids.cuh"
 
 
@@ -64,139 +57,6 @@ constexpr int default_block_size{512};
 
 
 #include "common/factorization/par_ilu_kernels.hpp.inc"
-
-
-template <typename ValueType, typename IndexType>
-void add_diagonal_elements(std::shared_ptr<const CudaExecutor> exec,
-                           matrix::Csr<ValueType, IndexType> *mtx,
-                           bool is_sorted)
-{
-    // TODO: Runtime can be optimized by choosing a appropriate size for the
-    //       subwarp dependent on the matrix properties
-    constexpr int subwarp_size = config::warp_size;
-    auto mtx_size = mtx->get_size();
-    auto num_rows = static_cast<IndexType>(mtx_size[0]);
-    auto num_cols = static_cast<IndexType>(mtx_size[1]);
-    size_type row_ptrs_size = num_rows + 1;
-
-    Array<IndexType> row_ptrs_addition(exec, row_ptrs_size);
-    Array<bool> needs_change_host{exec->get_master(), 1};
-    needs_change_host.get_data()[0] = false;
-    Array<bool> needs_change_device{exec, 1};
-    needs_change_device = needs_change_host;
-
-    auto cuda_old_values = as_cuda_type(mtx->get_const_values());
-    auto cuda_old_col_idxs = as_cuda_type(mtx->get_const_col_idxs());
-    auto cuda_old_row_ptrs = as_cuda_type(mtx->get_row_ptrs());
-    auto cuda_row_ptrs_add = as_cuda_type(row_ptrs_addition.get_data());
-
-    const dim3 block_dim{default_block_size, 1, 1};
-    const dim3 grid_dim{
-        static_cast<uint32>(ceildiv(num_rows, block_dim.x / subwarp_size)), 1,
-        1};
-    if (is_sorted) {
-        kernel::find_missing_diagonal_elements<true, subwarp_size>
-            <<<grid_dim, block_dim>>>(
-                num_rows, num_cols, cuda_old_col_idxs, cuda_old_row_ptrs,
-                cuda_row_ptrs_add,
-                as_cuda_type(needs_change_device.get_data()));
-    } else {
-        kernel::find_missing_diagonal_elements<false, subwarp_size>
-            <<<grid_dim, block_dim>>>(
-                num_rows, num_cols, cuda_old_col_idxs, cuda_old_row_ptrs,
-                cuda_row_ptrs_add,
-                as_cuda_type(needs_change_device.get_data()));
-    }
-    needs_change_host = needs_change_device;
-    if (!needs_change_host.get_const_data()[0]) {
-        return;
-    }
-
-    prefix_sum(exec, cuda_row_ptrs_add, row_ptrs_size);
-    exec->synchronize();
-
-    IndexType total_additions{};
-    exec->get_master()->copy_from(
-        exec.get(), 1, cuda_row_ptrs_add + row_ptrs_size - 1, &total_additions);
-    size_type new_num_elems = static_cast<size_type>(total_additions) +
-                              mtx->get_num_stored_elements();
-
-
-    Array<ValueType> new_values{exec, new_num_elems};
-    Array<IndexType> new_col_idxs{exec, new_num_elems};
-    auto cuda_new_values = as_cuda_type(new_values.get_data());
-    auto cuda_new_col_idxs = as_cuda_type(new_col_idxs.get_data());
-
-    kernel::add_missing_diagonal_elements<subwarp_size>
-        <<<grid_dim, block_dim>>>(num_rows, cuda_old_values, cuda_old_col_idxs,
-                                  cuda_old_row_ptrs, cuda_new_values,
-                                  cuda_new_col_idxs, cuda_row_ptrs_add);
-
-    const dim3 grid_dim_row_ptrs_update{
-        static_cast<uint32>(ceildiv(num_rows, block_dim.x)), 1, 1};
-    kernel::update_row_ptrs<<<grid_dim_row_ptrs_update, block_dim>>>(
-        num_rows + 1, cuda_old_row_ptrs, cuda_row_ptrs_add);
-
-    matrix::CsrBuilder<ValueType, IndexType> mtx_builder{mtx};
-    mtx_builder.get_value_array() = std::move(new_values);
-    mtx_builder.get_col_idx_array() = std::move(new_col_idxs);
-}
-
-GKO_INSTANTIATE_FOR_EACH_VALUE_AND_INDEX_TYPE(
-    GKO_DECLARE_PAR_ILU_ADD_DIAGONAL_ELEMENTS_KERNEL);
-
-
-template <typename ValueType, typename IndexType>
-void initialize_row_ptrs_l_u(
-    std::shared_ptr<const CudaExecutor> exec,
-    const matrix::Csr<ValueType, IndexType> *system_matrix,
-    IndexType *l_row_ptrs, IndexType *u_row_ptrs)
-{
-    const size_type num_rows{system_matrix->get_size()[0]};
-
-    const dim3 block_size{default_block_size, 1, 1};
-    const uint32 number_blocks =
-        ceildiv(num_rows, static_cast<size_type>(block_size.x));
-    const dim3 grid_dim{number_blocks, 1, 1};
-
-    kernel::count_nnz_per_l_u_row<<<grid_dim, block_size, 0, 0>>>(
-        num_rows, as_cuda_type(system_matrix->get_const_row_ptrs()),
-        as_cuda_type(system_matrix->get_const_col_idxs()),
-        as_cuda_type(system_matrix->get_const_values()),
-        as_cuda_type(l_row_ptrs), as_cuda_type(u_row_ptrs));
-
-    prefix_sum(exec, l_row_ptrs, num_rows + 1);
-    prefix_sum(exec, u_row_ptrs, num_rows + 1);
-}
-
-GKO_INSTANTIATE_FOR_EACH_VALUE_AND_INDEX_TYPE(
-    GKO_DECLARE_PAR_ILU_INITIALIZE_ROW_PTRS_L_U_KERNEL);
-
-
-template <typename ValueType, typename IndexType>
-void initialize_l_u(std::shared_ptr<const CudaExecutor> exec,
-                    const matrix::Csr<ValueType, IndexType> *system_matrix,
-                    matrix::Csr<ValueType, IndexType> *csr_l,
-                    matrix::Csr<ValueType, IndexType> *csr_u)
-{
-    const size_type num_rows{system_matrix->get_size()[0]};
-    const dim3 block_size{default_block_size, 1, 1};
-    const dim3 grid_dim{static_cast<uint32>(ceildiv(
-                            num_rows, static_cast<size_type>(block_size.x))),
-                        1, 1};
-
-    kernel::initialize_l_u<<<grid_dim, block_size, 0, 0>>>(
-        num_rows, as_cuda_type(system_matrix->get_const_row_ptrs()),
-        as_cuda_type(system_matrix->get_const_col_idxs()),
-        as_cuda_type(system_matrix->get_const_values()),
-        as_cuda_type(csr_l->get_const_row_ptrs()),
-        as_cuda_type(csr_l->get_col_idxs()), as_cuda_type(csr_l->get_values()),
-        as_cuda_type(csr_u->get_const_row_ptrs()),
-        as_cuda_type(csr_u->get_col_idxs()), as_cuda_type(csr_u->get_values()));
-}
-
-GKO_INSTANTIATE_FOR_EACH_VALUE_AND_INDEX_TYPE(
-    GKO_DECLARE_PAR_ILU_INITIALIZE_L_U_KERNEL);
 
 
 template <typename ValueType, typename IndexType>
