+++ omp/CMakeLists.txt
@@ -6,6 +6,7 @@ target_sources(ginkgo_omp
         base/version.cpp
         components/prefix_sum.cpp
         factorization/ilu_kernels.cpp
+        factorization/factorization_kernels.cpp
         factorization/par_ilu_kernels.cpp
         matrix/coo_kernels.cpp
         matrix/csr_kernels.cpp
