+++ reference/CMakeLists.txt
@@ -4,6 +4,7 @@ target_sources(ginkgo_reference
         base/version.cpp
         components/prefix_sum.cpp
         factorization/ilu_kernels.cpp
+        factorization/factorization_kernels.cpp
         factorization/par_ilu_kernels.cpp
         matrix/coo_kernels.cpp
         matrix/csr_kernels.cpp
